pyproject.toml
[tool.ruff]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]

select = ["E4", "E7", "E9", "F", "I", "UP", "C90", "B"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]

quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


docstring-code-format = false


docstring-code-line-length = "dynamic"


[project]
dependencies = [
    "Django==5.1.5",
    "djangorestframework==3.15.2",
    "django-rest-swagger==2.2.0",
    "drf-yasg==1.21.8",
    "drf-nested-routers==0.94.1",
    "ruff==0.9.6",
    "uv==0.5.31",
    "django-filter==24.3",
    "celery==5.4.0",
    "redis==5.2.1",
    "django-celery-beat==2.7.0"
]
